+++ store/quiz_solve.ts
@@ -1,3 +1,4 @@
+/* eslint-disable no-param-reassign */
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 
 // user스토어의 초기값을 설정
@@ -8,7 +9,7 @@ const initialState: SolveQuizSetType = {
   quizMaker: { nickname: '', profile_img: '' },
   quizSetThumbnail: '',
   answerList: [],
-  description:'',
+  description: '',
 };
 
 // ducks 패턴을 지원하기 위해 나온 함수가 createSlice.
